explode hypervisor.rs into smaller files
