+++ src/devices/mod.rs
@@ -7,8 +7,8 @@ use crate::devices::threads::SubscriberEventManager;
 use crate::devices::virtio::block::{self, BlockArgs};
 use crate::devices::virtio::console::{self, ConsoleArgs};
 use crate::devices::virtio::{CommonArgs, MmioConfig};
+use crate::kvm::hypervisor::ioregionfd::IoRegionFd;
 use crate::kvm::hypervisor::Hypervisor;
-use crate::kvm::hypervisor::IoRegionFd;
 use crate::kvm::PhysMemAllocator;
 use crate::result::Result;
 use crate::tracer::proc::Mapping;
