+++ src/devices/virtio/console/log_handler.rs
@@ -15,7 +15,7 @@ use vm_memory::Bytes;
 use vm_memory::{self, GuestAddressSpace};
 
 use crate::devices::virtio::SignalUsedQueue;
-use crate::kvm::hypervisor::IoEvent;
+use crate::kvm::hypervisor::ioevent::IoEvent;
 
 #[derive(Debug)]
 pub enum Error {
