+++ src/guest_mem.rs
@@ -1,7 +1,7 @@
 use std::ops::Range;
 use std::sync::Arc;
 
-use crate::kvm::hypervisor::PhysMem;
+use crate::kvm::hypervisor::memory::PhysMem;
 use crate::page_math::huge_page_size;
 use crate::page_table::{
     self, PageTable, PageTableFlags, PageTableIteratorValue, PhysAddr, VirtMem,
