+++ src/kvm/hypervisor/hypervisor.rs
@@ -1,135 +1,29 @@
 use crate::page_table::PhysAddr;
 use crate::tracer::inject_syscall;
 use kvm_bindings as kvmb;
-use libc::{c_int, c_void};
+use libc::c_int;
 use log::*;
-use nix::poll::{ppoll, PollFd, PollFlags};
-use nix::sys::signal::SigSet;
-use nix::sys::socket::{socketpair, AddressFamily, SockFlag, SockType};
-use nix::sys::time::TimeSpec;
 use nix::unistd::Pid;
-use nix::unistd::{close, read, write};
-use simple_error::{bail, simple_error, try_with};
+use simple_error::{bail, try_with};
 use std::ffi::OsStr;
 use std::marker::PhantomData;
 use std::mem::size_of;
-use std::mem::MaybeUninit;
 use std::os::unix::io::AsRawFd;
 use std::os::unix::prelude::RawFd;
 use std::sync::{Arc, Mutex, RwLock, RwLockWriteGuard};
-use std::time::Duration;
-use vm_memory::remote_mem;
 use vmm_sys_util::eventfd::{EventFd, EFD_NONBLOCK};
 
+use super::ioeventfd::IoEventFd;
+use super::ioregionfd::IoRegionFd;
+use super::memory::*;
 use crate::cpu;
-use crate::devices::virtio::{register_ioeventfd, MmioConfig};
-use crate::devices::USE_IOREGIONFD;
 use crate::kvm::fd_transfer;
 use crate::kvm::ioctls;
-use crate::kvm::ioctls::kvm_ioregion;
-use crate::kvm::ioctls::{ioregionfd_cmd, ioregionfd_resp};
 use crate::kvm::tracee::{kvm_msrs, Tracee};
 use crate::page_math::{self, compute_host_offset};
 use crate::result::Result;
 use crate::tracer::proc::{openpid, Mapping, PidHandle};
 use crate::tracer::wrap_syscall::KvmRunWrapper;
-
-pub fn process_read<T: Sized + Copy>(pid: Pid, addr: *const c_void) -> Result<T> {
-    remote_mem::process_read(pid, addr).map_err(|e| simple_error!("{}", e))
-}
-
-pub fn process_write<T: Sized + Copy>(pid: Pid, addr: *mut c_void, val: &T) -> Result<()> {
-    remote_mem::process_write(pid, addr, val).map_err(|e| simple_error!("{}", e))
-}
-
-/// Hypervisor Memory
-#[derive(Debug)]
-pub struct HvMem<T: Copy> {
-    pub ptr: libc::uintptr_t,
-    pid: Pid,
-    tracee: Arc<RwLock<Tracee>>,
-    phantom: PhantomData<T>,
-}
-
-impl<T: Copy> Drop for HvMem<T> {
-    fn drop(&mut self) {
-        // Useful for debugging
-        //warn!("SKIP CLEANUP");
-        //return;
-        let tracee = match self.tracee.write() {
-            Err(e) => {
-                warn!("Could not aquire lock to drop HvMem: {}", e);
-                return;
-            }
-            Ok(t) => t,
-        };
-        if let Err(e) = tracee.munmap(self.ptr as *mut c_void, size_of::<T>()) {
-            warn!("failed to unmap memory from process: {}", e);
-        }
-    }
-}
-
-impl<T: Copy> HvMem<T> {
-    pub fn read(&self) -> Result<T> {
-        process_read(self.pid, self.ptr as *mut c_void)
-    }
-    pub fn write(&self, val: &T) -> Result<()> {
-        process_write(self.pid, self.ptr as *mut c_void, val)
-    }
-}
-
-/// Physical Memory attached to a VM. Backed by `PhysMem.mem`.
-#[derive(Debug)]
-pub struct PhysMem<T: Copy> {
-    pub mem: HvMem<T>,
-    ioctl_arg: HvMem<kvmb::kvm_userspace_memory_region>,
-    pub guest_phys_addr: PhysAddr,
-}
-
-impl<T: Copy> Drop for PhysMem<T> {
-    fn drop(&mut self) {
-        // useful for debugging
-        //warn!("SKIP CLEANUP");
-        //return;
-
-        let tracee = match self.mem.tracee.write() {
-            Err(e) => {
-                warn!("Could not aquire lock to drop HvMem: {}", e);
-                return;
-            }
-            Ok(t) => t,
-        };
-        let mut ioctl_arg = match self.ioctl_arg.read() {
-            Err(e) => {
-                warn!("Could not read Hypervisor Memory to drop HvMem: {}", e);
-                return;
-            }
-            Ok(t) => t,
-        };
-        ioctl_arg.memory_size = 0; // indicates request for deletion
-        match self.ioctl_arg.write(&ioctl_arg) {
-            Err(e) => {
-                warn!("Could not write to Hypervisor Memory to drop HvMem: {}", e);
-                return;
-            }
-            Ok(t) => t,
-        };
-        let ret =
-            match tracee.vm_ioctl_with_ref(ioctls::KVM_SET_USER_MEMORY_REGION(), &self.ioctl_arg) {
-                Ok(ret) => ret,
-                Err(e) => {
-                    warn!("failed to remove memory from VM: {}", e);
-                    return;
-                }
-            };
-        if ret != 0 {
-            warn!(
-                "ioctl_with_ref to remove memory from VM returned error code: {}",
-                ret
-            )
-        }
-    }
-}
 
 #[allow(clippy::upper_case_acronyms)]
 pub struct VCPU {
@@ -146,481 +40,8 @@ pub struct Hypervisor {
     /// hypervisor memory where fd_num is mapped to.
     /// sorted by vcpu nr. TODO what if there exist cpu [0, 1, 4]?
     pub vcpu_maps: Vec<Mapping>,
-    tracee: Arc<RwLock<Tracee>>,
+    pub(super) tracee: Arc<RwLock<Tracee>>,
     pub wrapper: Mutex<Option<KvmRunWrapper>>,
-}
-
-/// Abstraction around IoEventFd and EventFd.
-/// They can't be placed into the same memory location without an enum, because
-/// vmm_sys_util::EventFd doesn't implement Sized which is required for `dyn`.
-pub enum IoEvent {
-    IoEventFd(IoEventFd),
-    EventFd(EventFd),
-}
-
-impl IoEvent {
-    pub fn register(
-        vmm: &Arc<Hypervisor>,
-        uioefd: &mut UserspaceIoEventFd,
-        mmio_cfg: &MmioConfig,
-        queue_idx: u64,
-    ) -> Result<IoEvent> {
-        if USE_IOREGIONFD {
-            let eventfd = try_with!(
-                uioefd.userpace_ioeventfd(Some(queue_idx as u32)),
-                "cannot register userspace ioeventfd"
-            );
-            Ok(IoEvent::EventFd(eventfd))
-        } else {
-            let ioeventfd = try_with!(
-                register_ioeventfd(vmm, mmio_cfg, queue_idx),
-                "cannot register ioeventfd"
-            );
-            Ok(IoEvent::IoEventFd(ioeventfd))
-        }
-    }
-}
-
-impl AsRawFd for IoEvent {
-    fn as_raw_fd(&self) -> RawFd {
-        match self {
-            IoEvent::IoEventFd(e) => e.as_raw_fd(),
-            IoEvent::EventFd(e) => e.as_raw_fd(),
-        }
-    }
-}
-
-impl Deref for IoEvent {
-    type Target = EventFd;
-    fn deref(&self) -> &EventFd {
-        match self {
-            IoEvent::IoEventFd(e) => e,
-            IoEvent::EventFd(e) => e,
-        }
-    }
-}
-
-pub struct IoEventFd {
-    fd: EventFd,
-    hv_eventfd: RawFd,
-    hv_mem: HvMem<kvmb::kvm_ioeventfd>,
-    tracee: Arc<RwLock<Tracee>>,
-    guest_addr: u64,
-    len: u32,
-    datamatch: Option<u64>,
-}
-
-fn kvm_ioeventfd(
-    hv_eventfd: RawFd,
-    guest_addr: u64,
-    len: u32,
-    datamatch: Option<u64>,
-) -> kvmb::kvm_ioeventfd {
-    let mut flags = 0;
-    let mut datam = 0;
-    if let Some(data) = datamatch {
-        flags = 1 << kvmb::kvm_ioeventfd_flag_nr_datamatch;
-        datam = data;
-    }
-
-    kvmb::kvm_ioeventfd {
-        len,
-        datamatch: datam,
-        addr: guest_addr,
-        fd: hv_eventfd.as_raw_fd(),
-        flags,
-        ..Default::default()
-    }
-}
-impl IoEventFd {
-    fn new(
-        hv: &Hypervisor,
-        guest_addr: u64,
-        len: u32,
-        datamatch: Option<u64>,
-    ) -> Result<IoEventFd> {
-        let eventfd = try_with!(EventFd::new(EFD_NONBLOCK), "cannot create event fd");
-        info!(
-            "ioeventfd {}, guest phys addr {:#x}",
-            eventfd.as_raw_fd(),
-            guest_addr
-        );
-        let hv_eventfd = hv.transfer(vec![eventfd.as_raw_fd()].as_slice())?[0];
-        let ioeventfd = kvm_ioeventfd(hv_eventfd, guest_addr, len, datamatch);
-
-        let mem = hv.alloc_mem()?;
-        mem.write(&ioeventfd)?;
-
-        let ret = {
-            let tracee = try_with!(
-                hv.tracee.read(),
-                "cannot obtain tracee read lock: poinsoned"
-            );
-            try_with!(
-                tracee.vm_ioctl_with_ref(ioctls::KVM_IOEVENTFD(), &mem),
-                "kvm ioeventfd ioctl injection failed"
-            )
-        };
-        if ret != 0 {
-            bail!("cannot register KVM_IOEVENTFD via ioctl: {:?}", ret);
-        }
-
-        Ok(IoEventFd {
-            guest_addr,
-            len,
-            datamatch,
-            hv_eventfd,
-            hv_mem: mem,
-            fd: eventfd,
-            tracee: hv.tracee.clone(),
-        })
-    }
-}
-
-impl Drop for IoEventFd {
-    fn drop(&mut self) {
-        let tracee = match self.tracee.read() {
-            Err(e) => {
-                warn!("IoEventfd: Could not aquire lock: {}", e);
-                return;
-            }
-            Ok(t) => t,
-        };
-        let mut ioeventfd =
-            kvm_ioeventfd(self.hv_eventfd, self.guest_addr, self.len, self.datamatch);
-        ioeventfd.flags |= 1 << kvmb::kvm_ioeventfd_flag_nr_deassign;
-
-        if let Err(e) = self.hv_mem.write(&ioeventfd) {
-            warn!(
-                "IoEventFd: Could not write to HvMem while dropping IoEventFd: {}",
-                e
-            );
-            return;
-        }
-
-        if let Err(e) = tracee.vm_ioctl_with_ref(ioctls::KVM_IOEVENTFD(), &self.hv_mem) {
-            warn!("IoEventfd: kvm ioeventfd ioctl injection failed: {}", e)
-        }
-
-        if let Err(e) = tracee.close(self.hv_eventfd) {
-            warn!("IoEventfd: failed to close eventfd in hypervisor: {}", e)
-        }
-    }
-}
-
-use std::ops::Deref;
-impl Deref for IoEventFd {
-    type Target = EventFd;
-    fn deref(&self) -> &EventFd {
-        &self.fd
-    }
-}
-
-impl AsRawFd for IoEventFd {
-    fn as_raw_fd(&self) -> RawFd {
-        self.fd.as_raw_fd()
-    }
-}
-
-struct UIoEFd {
-    datamatch: Option<u32>,
-    fd: EventFd,
-}
-
-/// For 32bit wide accesses on the queue notify register of virtio devices. Requires one UIoEventFd
-/// per device.
-pub struct UserspaceIoEventFd {
-    ioeventfds: Vec<UIoEFd>,
-}
-
-impl Default for UserspaceIoEventFd {
-    fn default() -> Self {
-        Self { ioeventfds: vec![] }
-    }
-}
-
-impl UserspaceIoEventFd {
-    pub fn userpace_ioeventfd(&mut self, datamatch: Option<u32>) -> Result<EventFd> {
-        if datamatch.is_none() && !self.ioeventfds.is_empty() {
-            bail!("cannot add a userspace ioeventfd without datamatch when others with datamatch have already been registered");
-        }
-
-        let fd = try_with!(
-            EventFd::new(EFD_NONBLOCK),
-            "cannot create non-blocking eventfd for uioefd"
-        );
-        log::info!("eventfd {:?} for ioregionfd", fd.as_raw_fd(),);
-        let uioefd = UIoEFd {
-            datamatch,
-            fd: try_with!(fd.try_clone(), "cannot clone uioefd"),
-        };
-        self.ioeventfds.push(uioefd);
-        Ok(fd)
-    }
-
-    /// Callback for writes to QueueNotify register. Forwards notification to the corresponding
-    /// EventFd reader.
-    pub fn queue_notify(&self, val: u32) {
-        let criterion = |uioefd: &&UIoEFd| match uioefd.datamatch {
-            Some(dm) => dm == val,
-            None => true,
-        };
-        if let Some(ioefd) = self.ioeventfds.iter().find(criterion) {
-            // ioefd is the correct fd from our list
-            if let Err(e) = ioefd.fd.write(1) {
-                log::trace!(
-                    "cannot write to UserspaceIoEventFd (datamatch {:?}, fd {}): {}",
-                    ioefd.datamatch,
-                    ioefd.fd.as_raw_fd(),
-                    e
-                );
-            }
-        } else {
-            log::trace!("cannot find datamatch for ");
-        }
-    }
-}
-
-/// Implements the KVM IoRegionFd feature.
-pub struct IoRegionFd {
-    tracee: Arc<RwLock<Tracee>>,
-    hv_mem: HvMem<kvm_ioregion>,
-    ioregion: kvm_ioregion,
-    rfile: RawFd, // our end: we write responses here
-    wfile: RawFd, // we read commands from here
-    rf_hv: RawFd, // their end: will be transferred to hyperisor
-    wf_hv: RawFd,
-    hv_rf_hv: RawFd, // rf_hv, but in hypervisor process
-    hv_wf_hv: RawFd,
-}
-
-impl IoRegionFd {
-    fn new(hv: &Hypervisor, guest_paddr: u64, len: usize) -> Result<Self> {
-        if Self::capability_present(hv)? {
-            bail!("This operation requires KVM_CAP_IOREGIONFD which your KVM does not have.");
-        }
-
-        let (rf_dev, rf_hv) = try_with!(
-            socketpair(
-                AddressFamily::Unix,
-                SockType::SeqPacket,
-                None,
-                SockFlag::SOCK_CLOEXEC
-            ),
-            "cannot create sockets for IoRegionFd"
-        );
-        let (wf_dev, wf_hv) = try_with!(
-            socketpair(
-                AddressFamily::Unix,
-                SockType::SeqPacket,
-                None,
-                SockFlag::SOCK_CLOEXEC
-            ),
-            "cannot create sockets for IoRegionFd"
-        );
-        let hv_rf_hv = hv.transfer(vec![rf_hv.as_raw_fd()].as_slice())?[0];
-        let hv_wf_hv = hv.transfer(vec![wf_hv.as_raw_fd()].as_slice())?[0];
-        let ioregion = kvm_ioregion::new(guest_paddr, len, hv_rf_hv, hv_wf_hv);
-        let mem = hv.alloc_mem()?;
-        mem.write(&ioregion)?;
-
-        let ret = {
-            let tracee = try_with!(
-                hv.tracee.read(),
-                "cannot obtain tracee read lock: poinsoned"
-            );
-            try_with!(
-                tracee.vm_ioctl_with_ref(ioctls::KVM_SET_IOREGION(), &mem),
-                "kvm ioeventfd ioctl injection failed"
-            )
-        };
-        if ret != 0 {
-            bail!("ioregionfd ioctl failed with {}", ret);
-        }
-
-        Ok(IoRegionFd {
-            tracee: hv.tracee.clone(),
-            hv_mem: mem,
-            ioregion,
-            rfile: rf_dev,
-            wfile: wf_dev,
-            rf_hv,
-            wf_hv,
-            hv_rf_hv,
-            hv_wf_hv,
-        })
-    }
-
-    pub fn fdclone(&mut self) -> RawIoRegionFd {
-        let pollfds = vec![PollFd::new(self.wfile, PollFlags::POLLIN)];
-        RawIoRegionFd {
-            rfile: self.rfile,
-            wfile: self.wfile,
-            ioregion: self.ioregion,
-            pollfds,
-        }
-    }
-
-    pub fn capability_present(hv: &Hypervisor) -> Result<bool> {
-        let has_cap = try_with!(
-            hv.check_extension(ioctls::KVM_CAP_IOREGIONFD as i32),
-            "cannot check kvm extension capabilities"
-        );
-        Ok(has_cap == 0)
-    }
-}
-
-impl Drop for IoRegionFd {
-    fn drop(&mut self) {
-        let tracee = match self.tracee.read() {
-            Err(e) => {
-                warn!("IoEventfd: Could not aquire lock: {}", e);
-                return;
-            }
-            Ok(t) => t,
-        };
-
-        let mut ioregion = self.ioregion;
-        ioregion.rfd = -1;
-        ioregion.wfd = -1;
-
-        if let Err(e) = self.hv_mem.write(&ioregion) {
-            warn!(
-                "IoRegionFd: Could not write to HvMem while dropping IoRegionFd: {}",
-                e
-            );
-            return;
-        }
-
-        match tracee.vm_ioctl_with_ref(ioctls::KVM_SET_IOREGION(), &self.hv_mem) {
-            Err(e) => warn!("IoRegionFd: kvm ioregionfd ioctl injection failed: {}", e),
-            Ok(ret) => {
-                if ret != 0 {
-                    warn!("IoRegionFd: kvm ioregionfd remove syscall failed: {}", ret);
-                }
-            }
-        }
-
-        match tracee.close(self.hv_rf_hv) {
-            Err(e) => warn!("IoRegionFd: close injection failed: {}", e),
-            Ok(ret) => {
-                if ret != 0 {
-                    warn!(
-                        "IoRegionFd: failed to close hv_rf_hv in hypervisor: {}",
-                        ret
-                    )
-                }
-            }
-        }
-
-        match tracee.close(self.hv_wf_hv) {
-            Err(e) => warn!("IoRegionFd: close injection failed: {}", e),
-            Ok(ret) => {
-                if ret != 0 {
-                    warn!(
-                        "IoRegionFd: failed to close hv_wf_hv in hypervisor: {}",
-                        ret
-                    )
-                }
-            }
-        }
-
-        if let Err(e) = close(self.rf_hv) {
-            warn!("IoRegionFd: failed to close rf_hv: {}", e)
-        }
-
-        if let Err(e) = close(self.wf_hv) {
-            warn!("IoRegionFd: failed to close wf_hv: {}", e)
-        }
-
-        if let Err(e) = close(self.rfile) {
-            warn!("IoRegionFd: failed to close rfile: {}", e)
-        }
-
-        if let Err(e) = close(self.wfile) {
-            warn!("IoRegionFd: failed to close wfile: {}", e)
-        }
-    }
-}
-
-/// A handle implementing clone, read and write for the non-clonable IoRegionFd.
-#[derive(Debug, Clone)]
-pub struct RawIoRegionFd {
-    rfile: RawFd, // our end: we write responses here
-    wfile: RawFd, // we read commands from here
-    pollfds: Vec<PollFd>,
-    pub ioregion: kvm_ioregion,
-}
-
-impl RawIoRegionFd {
-    /// receive read and write events/commands
-    pub fn read(&mut self) -> Result<Option<ioregionfd_cmd>> {
-        let len = size_of::<ioregionfd_cmd>();
-        let mut t_mem = MaybeUninit::<ioregionfd_cmd>::uninit();
-        // safe, because slice.len() == len
-        let t_slice = unsafe { std::slice::from_raw_parts_mut(t_mem.as_mut_ptr() as *mut u8, len) };
-
-        // read
-        let timeout = TimeSpec::from(Duration::from_millis(300));
-        let nr_events = try_with!(
-            ppoll(&mut self.pollfds, Some(timeout), SigSet::empty()),
-            "read/ppoll failed"
-        );
-        if nr_events == 0 || self.pollfds[0].revents().is_none() {
-            return Ok(None);
-        }
-        let read = try_with!(
-            read(self.wfile, t_slice),
-            "read on ioregionfd {} failed",
-            self.wfile
-        );
-        if read != len {
-            bail!(
-                "read returned ioregionfd command of incorrect length {}",
-                read
-            );
-        }
-
-        // safe, because we wrote exactly the correct amount of bytes (len)
-        let cmd: ioregionfd_cmd = unsafe { t_mem.assume_init() };
-        log::trace!(
-            "read {:?}, {:?}, response={}: {:?}",
-            cmd.info.cmd(),
-            cmd.info.size(),
-            cmd.info.is_response(),
-            cmd
-        );
-        Ok(Some(cmd))
-    }
-
-    /// Write a response back to the VM.
-    pub fn write_slice(&self, data: &[u8]) -> Result<()> {
-        log::trace!("write_slice()");
-        let mut arr = [0u8; 8];
-        let arr_slice = &mut arr[0..data.len()];
-        arr_slice.copy_from_slice(data);
-        self.write(u64::from_ne_bytes(arr))
-    }
-
-    /// Write a response back to the VM.
-    pub fn write(&self, data: u64) -> Result<()> {
-        log::trace!("write {:x}", data);
-        let len = size_of::<ioregionfd_resp>();
-        let response = ioregionfd_resp::new(data);
-        // safe, because we won't need t_bytes for longer than this stack frame
-        let t_bytes = unsafe {
-            std::slice::from_raw_parts((&response as *const ioregionfd_resp) as *const u8, len)
-        };
-        let written = try_with!(
-            write(self.rfile, t_bytes),
-            "write on ioregionfd {} failed",
-            self.rfile
-        );
-        if written != len {
-            bail!("cannot write entire ioregionfd command {}/{}", written, len);
-        }
-        Ok(())
-    }
 }
 
 impl Hypervisor {
