+++ src/kvm/hypervisor/mod.rs
@@ -1,4 +1,9 @@
 #[allow(clippy::module_inception)]
 pub mod hypervisor;
+pub mod ioevent;
+pub mod ioeventfd;
+pub mod ioregionfd;
+pub mod memory;
+pub mod userspaceioeventfd;
 
 pub use self::hypervisor::*;
